Comitting after adding more in depth main menu functionality.
I have orgonised my menu lists and also grouped my switch cases for the main menu into functions for easier code readability. The load game function however needs to be added but all the other options work.

I have also added the option to change the speed at which the text scrolls in all of the game with 3 predefined settings "slow", "medium" and "fast". These read and write to the config.txt file in the project folder. I have added functionality so that it doesn't matter which line it is on in the file but everytime the user changes it in game it creates a new file each time to reduce risk of corruption or user input error. The file contains the strings slow medium and fast and the program interprets that to change the speed in game.