commiting after exploring how to do file save game logic. Decided in essence of time to use a similar approach as to how the config file is serialised to store each piece of data in the save game.
each object needing to saved will be serialised and then adding to a corosponding key in the dictionary.

The plan is each of these can be loaded from the save file as needed and default values like room name and descriptions do not needed to be saved.

The player attributes will also be saved in the same fashion as this.

When a new game is created, a fresh savegame file will be created and new objects of each class saved with default values. This can be scaled up as the game project grows.

I have also created a few other classes like storage and item. the storage class I have added two trials methods for the player removing and placing an item in them.