modified display text function so when user presses enter or space it now just changes the scroll speed to 0 rather than using the substring method as I have added an auto text wrap function to allow strings to be passed to it and have them autowrap to fit the length of the console box.
I have been experimenting with this concept using the opening.txt file. with the substring method, when the user pressed space or enter it didn't wrap the final bit of text that the user was trying to skip the animation for so the above change was made. My plan is to make it so that in the txt file it doesn't matter about the paragraph structure, it will just display next text only when it sees a new line.