comitting after adding more dialogue to the game and several functions modified and added.
I have created various display text functions are for instance the text that is displayed above giving the player a bunch of options doesn't need a console.readkey afterwards.

My plan is that each room will have its own cs file containing all of the code needed for each. I plan to load details for instance if a player has entered that room before in the JSON save file.