comitting after adding extra functionality to the room 1 dialogue. I have wrapped the main menu in a function so for instance speaking to Dave can be repeated over and over and the case just simply calls the main menu function.
There is a bug at the moment where whenever a different person talks in diagloue it causes the artwork to flash as it clears the console and reloads it. The solution would be to use the cursor to change the name of the person talking rather than doing a full console clear everytime. I may change the display artwork functions in the future to accomedate this.