comitting after fixing the bug where whenever the person talking changed in the dialogue it would cause the image to flash. I got around this by creating a new function called change name that instead of needing to clear the full console everytime the person talking updates, it now just moves the cursor position. I have amended this also in my read text from file function and removed the continue statement that was causing me problems trying to debug it.
All of the known bugs have been resolved and I now play on making it so a weapon has to be found before the goblin will let them pass out of the room. I plan on having a check that looks for the sword in the players inventory and if they do they will have the option to brandish it to the goblin who will then let them pass.

I also need to work on my save game functionality. My plan as of now is that whenever a player enters a new room they have the option to save the state of the game.