comitting after making progress with the room 1 options. I have set all the class fields to private and added methods to modify and get the information.
I have added extra attributes to the storage class so for instance if the player has previously interacted with it and I also plan to add a marker system to identify whch room the player is in for when I do the save and load function. Currently I am working on storage and player inventories. I have a function for each to both add and remove items. These will both be JSON serialised for when the game saves.