comitting after finally getting the same game functionality working. The save game feature I realised wasn't working because I wasn't JSONIncluding the attributes of the Item class so it had nothing to add when it serialised it. I have also modified my transfer item methods of the storage so now transfering to a player is done via a seperate method to avoid confusion and to improve code quality.
I have also added the functionality to select a savegame file from a list of filenames. currently I am splitting the strings of the savegame file names and just displaying the character name to the user without the filepath or .JSON on the end. as the order in which they will be displayed to the user will be their index in the list, I am using this with the output of the menu function to select the correct save game file.

If time permitts I would like to add a check to prevent someone overriding the save file game name with a new save but I may run out of time to add this feature.