fixed a bug in the Game Saves cs file where whenever you would click load a game and then return to the main menu and click newgame it would crash and give an exception that "return to main menu .Json not found"
I discovered this was due to me missing an else statement around the code that handles the logic that loads the file. I initially I though becuase in the if statement it calls a function that the rest of the code in the old function wouldn't ge excecuted but a simple else statmeent around the 3 lines of code has fixed the problem.