comitting after a rather lengthy session.
I have added a game death screen that will occour if the player's health has dropped to 0. This screen will display 3 options that will allow them to start a new game, load from a save or go to the main menu.

I have also added a feature that checks if the player has the sword in their inventory. If they do it allows them to brandish it to the goblin and the goblin will then pass. A lot of this was fairly easy as I was reusing code from my previous functions. I am currently writing the code for the final combat scene that will happen in the next room.