comitting after finishing for the day. Added the final combat scene that needs finishing.
Currently the code doesn't work and I was having issues with the damage and health flipping between float and int values. for simplicity I have opted to add the random damge increases or health potion random increase to just be a simple intager value.

To finish I just need to conclude the turns of the combat. I deally I wanted to combat to be resuable functions for if there were to be any future development and for good code quality.