finished the final combat scene off. Converted all the floats to ints for the combat calculations and added some sound effects to it. I have also added a few methods that reset everything back to its starting point if the play selects a new game.
In retrospect, it would have been easier to use constructors rather than keep using the same class objects and just resetting them back to their origional poistion everytime. I just need to add a few final sound fx and also add all my files to the file load checker.

Currently the load game feature doesn't fully work as it has the problem of not being able to recognise the class objects contained in the player inventory or the chest once they are deserialised and I am not sure how to resolve this.